requirements.txt
amazon-braket-sdk>=1.71.0
amazon-braket-default-simulator>=1.20.1
numpy>=1.24.0
qiskit>=1.0.0
qiskit-ionq>=0.5.0
pylatexenc>=2.10
python-dotenv>=1.0.0
